Moved class-specific methods to ClassDumper
